+++ HISTORY.md
@@ -43,6 +43,7 @@
 ### 2023-05-01
 * objects: changed new costume name function to reflect the case-sensitivity preference setting
 * objects, threads: fixed costume naming rule for video snaps
+* objects: changed new sound name function to reflect the case-sensitivity preference setting
 
 ### 2023-05-01
 * blocks, objects: changed messages dropdown menu to reflect the case-sensitivity preference setting